+++ mkdocs.yml
@@ -97,4 +97,4 @@ extra_javascript:
 watch:
   - qadence-measurements
   - qadence-mitigations
-  - qadence-libs
+  - qadence-gradient
