+++ qadence-gradient/docs/javascripts/mathjax.js
