+++ qadence-gradient/pyproject.toml
@@ -3,8 +3,8 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "qadence-libs"
-description = "Libraries to extend Qadence functionalities."
+name = "qadence-gradient"
+description = "Gradient libraries to extend Qadence functionalities."
 readme = "README.md"
 authors = [
     { name = "Mario Dagrada", email = "mario.dagrada@pasqal.com" },
@@ -15,7 +15,7 @@ authors = [
     { name = "Niklas Heim", email = "niklas.heim@pasqal.com" },
     { name = "Ignacio Fernández Graña", email = "ignacio.fernandez-grana@pasqal.com" },
 ]
-requires-python = ">=3.9,<3.12"
+requires-python = ">=3.9,<3.14"
 license = {text = "Apache 2.0"}
 keywords = ["quantum"]
 version = "0.1.3"
@@ -26,6 +26,8 @@ classifiers=[
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Programming Language :: Python :: Implementation :: CPython",
     "Programming Language :: Python :: Implementation :: PyPy",
 ]
@@ -62,14 +64,14 @@ dependencies = [
 features = ["visualization"]
 
 [tool.hatch.envs.default.scripts]
-test = "pytest -n auto --cov-report=xml --cov-config=pyproject.toml --cov=qadence_libs --cov=tests --ignore=./tests/test_examples.py {args}"
+test = "pytest -n auto --cov-report=xml --cov-config=pyproject.toml --cov=qadence_gradient --cov=tests --ignore=./tests/test_examples.py {args}"
 test-examples = "pytest ./tests/test_examples.py {args}"
 no-cov = "cov --no-cov {args}"
 test-docs = "mkdocs build --clean --strict"
 test-all = "pytest -n auto {args}  && mkdocs build --clean --strict"
 
 [tool.pytest.ini_options]
-pythonpath = ["qadence_libs"]
+pythonpath = ["qadence_gradient"]
 testpaths = ["tests"]
 addopts = """-vvv"""
 xfail_strict = true
@@ -117,7 +119,7 @@ exclude = [
 ]
 
 [tool.hatch.build.targets.wheel]
-packages = ["qadence_libs"]
+packages = ["qadence_gradient"]
 
 [tool.coverage.run]
 branch = true
