+++ qadence-gradient/qadence_gradient/constructors/hea.py
@@ -50,7 +50,7 @@ def hea(
     Examples:
     ```python exec="on" source="material-block" result="json"
     from qadence import RZ, RX
-    from qadence_libs import hea
+    from qadence_gradient import hea
 
     # create the circuit
     n_qubits, depth = 2, 4
