+++ qadence-gradient/tests/constructors/test_feature_maps.py
@@ -5,13 +5,13 @@ from typing import Callable
 import pytest
 import sympy
 import torch
-from metrics_libs import ATOL_64
+from metrics_gradient import ATOL_64
 from qadence.execution import expectation, run
 from qadence.operations import PHASE, RX, X, Z
 from qadence.parameters import FeatureParameter
 from qadence.types import PI, BasisSet, ReuploadScaling
 
-from qadence_libs.constructors import exp_fourier_feature_map, feature_map
+from qadence_gradient.constructors import exp_fourier_feature_map, feature_map
 
 PARAM_DICT_0 = {
     "support": None,
