+++ qadence-gradient/tests/qinfo_tools/test_spsa.py
@@ -8,7 +8,7 @@ import torch
 from qadence import QNN, Overlap
 from torch import Size
 
-from qadence_libs.qinfo_tools.spsa import _shifted_overlap, spsa_2gradient_step
+from qadence_gradient.qinfo_tools.spsa import _shifted_overlap, spsa_2gradient_step
 
 SEED = 42
 torch.manual_seed(SEED)
